Заменил int на size_t в функции spiral

+++ pozdeev.andrey/P3/pozdeev.hpp
@@ -10,7 +10,7 @@ namespace pozdeev {
   constexpr int ERROR_INVALID_VALUE = 92;
 
   int countNonSequentialRows(const int* matrix, size_t rows, size_t cols);
-  void spiral(int* matrix, int rows, int cols);
+  void spiral(int* matrix, size_t rows, size_t cols);
   int numAnalysis(const char* s);
   int processStatic(std::ifstream& fin, std::ofstream& fout, size_t rows, size_t cols);
   int processDynamic(std::ifstream& fin, std::ofstream& fout, size_t rows, size_t cols);
